Server/requirements.txt
Flask>=2.2
Flask-SQLAlchemy==3.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
Flask-CORS==3.0.10